MNT: Move interface.utils.eval_results to interface(.base)
This is solely an internal code move. Fixes #6694
